Capital evolution with monthly savings
Capital evolution takes a combination of a linear model for the part one can control (initial capital and monthly capital increase) as well as an exponential model to account for both interest rates as well as inflation. An example of it was calculated for the following parameters : a savings account with 1000 Euro initial capital, 100 Euro saving rate per month, 6 percent interest rate (typical for ETFs) and 2 percent inflation. This example can be seen in the figure development_of_total_capital.png